Add column names for MCMC and MC3 results
Now the chain matrices returned by ggumMCMC and ggumMC3 have the column for
each parameter named. The first n columns will be named thetai for i in
{1, 2, ..., n} (the theta parameter for each respondent). The next m columns
will be named alphaj for j in {1, 2, ..., m} (the alpha parameter for each
item), and similarly with the next m columns and deltaj. After that, the
columns follow the naming convention tauj_k, identifying both the item and
the option. Additionally, beginning with this commit, ggumMCMC and ggumMC3
no longer return columns of 0 for the tauj_0 parameters; there's just not
much reason to increase the object size to store basically zero information.