Add ICC plotting function and color palettes
The item characteristic curve (ICC) funtion is often described as the
probability of a correct answer as a function of respondents' latent ability.
In the context of the GGUM, the ICC describes the expected value of response
given respondents' latent trait (Roberts and Thompson 2011). So, the icc()
function produces a plot of expected response as a function of theta.

Also here I separate out the definition of color palettes since now they will
be used by multiple functions.