Specify Travis CI build matrix
Until https://github.com/travis-ci/travis-ci/issues/10295 is fixed,
OSX builds will fail if fortran is true, but Linux builds will fail if
fortran is false, since RcppDist has to be built from source on Linux.
So, I'm trying to set up a situation where fortran is false on the OSX
build and true on the Ubuntu build.