Allow selection of recorded chain in ggumMC3()
I added this functionality purely for the sake of a figure in the
paper demonstrating the 'melting' of the posterior surface. I'm
putting this branch with the functionality on GitHub for replication
purposes, but I don't want to merge it into master, at least for now.
It seems like it might be a bad idea to let the user do this, as
hotter chains aren't going to have the desired target density. On the
other hand, we may eventually bring this into master to let users
record hotter chains for inspection and tuning purposes.