Add examples and update vignette
I added examples for every exported function, though examples for every
function that would require sampling are marked as dontrun.
I also fixed a typo in the vignette regarding how we tune proposals;
it had said the increments/decrements were by 0.01, but that isn't true
as described in the documentation for tune_proposals().
Since the increment/decrement rule was slightly involved, I omitted it.