Begin documentation changes for CRAN resubmit
For acceptance to CRAN, we have to add a GGUM reference in the
DESCRIPTION file and replace dontrun by donttest. I've added the
reference. Instead of replacing dontrun by donttest, I changed the
examples to be small enough to run in < 5 sec. anyway. I also fixed
a few things in the docs to be a little more clear. I still need to
double check everything, but committing now because I'm changing local
machines.