Improve proposal tuning
I made two improvements to the proposal tuning code, one of which might
be considered a bug fix. First, I made the acceptance bound checks an if
else if check rather than two if checks (not sure how I missed that the
first go-round). Second, I found out it was technically possible to
accidentally end up with an s.d. of 0, so rather than checking for
negative values, we now check if the s.d. is below a very low number
(1e-6).