Prepare for CRAN resubmission
In R 4.0, matrices will also have class 'array'. To deal with this,
in tests, I checked the major version and expected matrices to be of
class 'matrix' and class 'array' if the major version was greater than
3 and of class 'matrix' otherwise. However, for now, apparently whether
this new behavior occurs depends on an environment variable, so we
ended up failing tests on a couple of CRAN platforms on devel. I put
in a fix and incremented the patch version of the package; will
resubmit if tests on win-builder, etc. all pass.